refactored the folder structure
