+++ lib/presentation/widgets/diary_card.dart
