+++ pubspec.yaml
@@ -23,6 +23,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_bloc: ^6.0.6
 
 
   # The following adds the Cupertino Icons font to your application.
